+++ examples/VibrationSensor/VibrationSensor.ino
@@ -37,7 +37,9 @@
 
 // Enable and select radio type attached
 #define MY_RADIO_NRF24
+//#define MY_RADIO_NRF5_ESB
 //#define MY_RADIO_RFM69
+//#define MY_RADIO_RFM95
 
 #include <MySensors.h>
 #include <Wire.h>
@@ -46,7 +48,7 @@
 #define VIBRATION_SENSOR_DIGITAL_PIN 3
 #define SensorLED     13
 
-unsigned long SLEEP_TIME = 10*1000; // Sleep time between reads (in seconds)
+uint32_t SLEEP_TIME = 10*1000; // Sleep time between reads (in seconds)
 
 //VARIABLES
 int val = 0;           // variable to store the value coming from the sensor
@@ -58,40 +60,42 @@ MyMessage vibrationMsg(CHILD_ID_VIBRATION, V_LEVEL);
 
 void setup()
 {
-  pinMode(VIBRATION_SENSOR_DIGITAL_PIN, INPUT);
-  attachInterrupt(digitalPinToInterrupt(VIBRATION_SENSOR_DIGITAL_PIN), blink, FALLING); // Trigger the blink function when the falling edge is detected
-  pinMode(SensorLED, OUTPUT);
+	pinMode(VIBRATION_SENSOR_DIGITAL_PIN, INPUT);
+	attachInterrupt(digitalPinToInterrupt(VIBRATION_SENSOR_DIGITAL_PIN), blink,
+	                FALLING); // Trigger the blink function when the falling edge is detected
+	pinMode(SensorLED, OUTPUT);
 }
 
-void presentation()  {
-  // Send the sketch version information to the gateway and Controller
-  sendSketchInfo("VIBRATION Sensor", "1.0");
+void presentation()
+{
+	// Send the sketch version information to the gateway and Controller
+	sendSketchInfo("VIBRATION Sensor", "1.0");
 
-  // Register all sensors to gateway (they will be created as child devices)
-  present(CHILD_ID_VIBRATION, S_VIBRATION);
+	// Register all sensors to gateway (they will be created as child devices)
+	present(CHILD_ID_VIBRATION, S_VIBRATION);
 }
 
 void loop()
 {
 
-  if(state>=40){ // basically below 40 so ignire basic level
-        send(vibrationMsg.set(int16_t(state)));
-        state = 0;
-        digitalWrite(SensorLED,HIGH);
-   }    else {
-        state = 0;
-        digitalWrite(SensorLED,LOW);
-  }
+	if(state>=40) { // basically below 40 so ignire basic level
+		send(vibrationMsg.set(int16_t(state)));
+		state = 0;
+		digitalWrite(SensorLED,HIGH);
+	}    else {
+		state = 0;
+		digitalWrite(SensorLED,LOW);
+	}
 
 
-  // Power down the radio.  Note that the radio will get powered back up
-  // on the next write() call.
-  delay(1000); //delay to allow serial to fully print before sleep
+	// Power down the radio.  Note that the radio will get powered back up
+	// on the next write() call.
+	delay(1000); //delay to allow serial to fully print before sleep
 
-  sleep(SLEEP_TIME); //sleep for: sleepTime
+	sleep(SLEEP_TIME); //sleep for: sleepTime
 }
 
 void blink()//Interrupts function
 {
-  state++;
+	state++;
 }
